Fixed several compiler warnings

+++ Clove/src/Clove/ImGui/ImGuiLayer.cpp
@@ -69,7 +69,7 @@ namespace clv{
 	void ImGuiLayer::end(){
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::get();
-		io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(app.getWindow().getWidth()), static_cast<float>(app.getWindow().getHeight()));
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
